+++ src/types.jl
@@ -38,6 +38,7 @@ mutable struct SolverOptions
     tabu_list_length                    :: Int64
     num_resolve_nlp_feasibility_pump    :: Int64
     mip_solver                          :: Union{Nothing, JuMP.OptimizerFactory}
+    allow_almost_solved                 :: Bool  
     allow_almost_solved_integral        :: Bool  
     registered_functions                :: Union{Nothing, Vector{RegisteredFunction}}
     
@@ -62,6 +63,10 @@ mutable struct JuniperProblem
     nl_solver_options   :: Vector{Tuple}
    
     model               :: JuMP.Model
+
+    relaxation_status   :: MOI.TerminationStatusCode
+    relaxation_objval   :: Float64
+    relaxation_solution :: Vector{Float64}
 
     status              :: MOI.TerminationStatusCode
     objval              :: Float64
@@ -153,8 +158,7 @@ end
 mutable struct Incumbent
     objval      :: Float64
     solution    :: Vector{Float64}
-    status      :: MOI.TerminationStatusCode
-    best_bound  :: Float64
+    only_almost :: Bool
 end
 
 mutable struct GainObj
@@ -170,6 +174,7 @@ end
 mutable struct BnBTreeObj
     m               :: Juniper.JuniperProblem
     incumbent       :: Incumbent
+    limit           :: Symbol
     obj_gain        :: GainObj
     disc2var_idx    :: Vector{Int64}
     var2disc_idx    :: Vector{Int64}
@@ -179,6 +184,7 @@ mutable struct BnBTreeObj
     nsolutions      :: Int64
     branch_nodes    :: Vector{BnBNode}
     best_bound      :: Float64
+    global_solver   :: Bool
 
     BnBTreeObj() = new()
 end